requirements.txt
celery==5.2.7
cryptography==40.0.1
Django==4.1.7
django-allow-cidr==0.6.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==20.1.0
openpyxl==3.1.2
psycopg2-binary==2.9.5
python-decouple==3.8
requests==2.28.2
sentry-sdk==1.17.0